requirements.txt
# Django y servidor WSGI
Django==4.2.20
gunicorn==23.0.0

# Async (si usas vistas async)
aiohttp==3.9.3
aiohttp-cors==0.7.0
asgiref==3.8.1

# Django REST Framework y herramientas relacionadas
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0

# Autenticación y entorno
python-dotenv==1.1.0
PyJWT==2.4.0

# Base de datos y conexión por URL
psycopg2==2.9.10
dj-database-url==2.3.0

# Utilidades
requests==2.27.1
python-slugify==5.0.2

# Redis (opcional, si usas caché o Celery)
redis==5.0.3

# Almacenamiento externo (S3, GCP, etc.)
boto3==1.34.92
botocore==1.34.92
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-storage==2.16.0